requirements.txt
aiohttp>=3.8.0
opencv-python>=4.5.0
Pillow>=9.0.0
PyYAML>=6.0
numpy>=1.21.0